use 24bit deviceid as uid instead of proprietary 8bit

+++ common/rfm69.h
@@ -24,25 +24,25 @@ class RFM69 {
 
     typedef struct {
       uint8_t message[RFM69_MAX_DATA_LEN+1];
-      uint8_t from;
+      uint32_t from;
       int16_t rssi;
     } Packet;
 
     RFM69(pins_t pin_cs, pins_t pin_interrupt);
     RFM69();
-    uint8_t  init(uint16_t freq, uint8_t node_id, uint8_t network_id);
-    uint8_t  init(uint8_t node_id, uint8_t network_id);
+    uint8_t  init(uint16_t freq, uint32_t node_id, uint8_t network_id);
+    uint8_t  init(uint32_t node_id, uint8_t network_id);
     uint8_t  read_reg(uint8_t addr);
     void     write_reg(uint8_t addr, uint8_t value);
     void     encrypt(const char* key);
     void     set_high_power();
     void     set_network(uint8_t id);
     void     set_power_level(uint8_t level);
-    uint8_t  send(uint8_t to, const void* buffer, uint8_t buffer_len, RFM69::Packet *response = 0);
-    uint8_t  send_retry(uint8_t to, const void* buffer, uint8_t buffer_len, RFM69::Packet *response, uint8_t retries);
+    uint8_t  send(uint32_t to, const void* buffer, uint8_t buffer_len, RFM69::Packet *response = 0);
+    uint8_t  send_retry(uint32_t to, const void* buffer, uint8_t buffer_len, RFM69::Packet *response, uint8_t retries);
     uint8_t  listen(RFM69::Packet *response, uint8_t timeout_enabled = 1);
     int16_t  read_rssi(uint8_t force = 0);
-    void     send_frame(uint8_t to, const void* buffer, uint8_t size, uint8_t request_ack, uint8_t send_ack);
+    void     send_frame(uint32_t to, const void* buffer, uint8_t size, uint8_t request_ack, uint8_t send_ack);
     void     sleep();
 
     static RFM69 *rfm69_ptr;
@@ -53,7 +53,7 @@ class RFM69 {
     Mode mode;
     pins_t pin_cs;
     uint8_t power_level;
-    uint8_t node_id;
+    uint32_t node_id;
     uint8_t spy_mode;
     void    init_vars(pins_t pins_cs, pins_t pin_interrupt);
